Handle Boost.System and adjust the code to handle the binary dependencies.
Boost.Filesystem (and the libraries depending on it) now requires
Boost.System at link time.  Boost.System was introduced in Boost 1.35 so
we must only try to find it if the version at hand is >= 1.35.

Bug reported by Daniel Herring with some code contributions of
Daniel Herring and Paul Balomiri.

Signed-off-by: Benoit Sigoure <tsuna@lrde.epita.fr>