Do not add useless -I for boost.
Currently these macros can add -I to CPPFLAGS although none is needed, and
none were required by the user.  This clutters compiler invocations.

The code seems to be ready to handle the case where Boost is found
in a default location (as it explicitly tries `for boost_dir in
"$with_boost/include" '' ...'), yet it is not functional because the code
wanted to read boost/version.hpp so it required to know its path.

The logic is now changed: we do *not* require to find the path to Boost
ourselves, after all this is not what will happen with the compiler which
will use *all* its -I to find it, but we run cpp to get the value we are
interested in.

	* build-aux/boost.m4 (_BOOST_EGREP_SED): New.
	(BOOST_REQUIRE): Do not require to find the file, just let the
	compiler/preprocessor do its job.