Installs in $boost_dir/boost-x_xx take precedent over $boost_dir
I hit a snag in the following scenario using boost.m4:

1) One copy of Boost is installed in /usr so that /usr/include/boost/version.hpp exists.
2) I wish to use a second copy of Boost installed elsewhere using
   --with-boost=/somedir, and this copy version.hpp located at
   /somedir/include/boost-x_xx/boost/version.hpp where boost-x_xx contains a
   version string.  This copy is installed so that the libraries are at
   /somedir/lib.

In this case, boost.m4 sets BOOST_CPPFLAGS to be -I/somedir instead of
detecting that -I/somedir/boost-x_xx is appropriate.  As a result,
compiling files with #include <boost/something> causes the
/usr/include/boost headers to used by the preprocessor but /somedir/lib
to be used by the linker.

boost.m4 attempted to address this using a '--layout=system' inner loop
like
    # Without --layout=system, Boost (or at least some versions) installs
    # itself in <prefix>/include/boost-<version>.  This inner loop helps to
    # find headers in such directories.
    # ...
    for boost_inc in "$boost_dir" "$boost_dir"/boost-*
    do
        # ...
    done
which does not handle my use case because during the $boost_dir portion
of the loop the headers are detected from
/usr/include/boost/version.hpp.

The fix is to change the order of the loop so that it looks for
--layout=system installs first, i.e. getting to
/somedir/include/boost-x_xx/boost/version.hpp before
/usr/include/boost/version.hpp.

Moreover, we'd like boost-y_yy to take precedence over boost-x_xx if
y_yy is a later version than x_xx.  This requires searching in 'sort by
version order' which is done using 'sort -rn -t- -k2'

This commit fixes the issue.  It does, however, bring in a dependency on
ls, grep, and sort.  All of these are used (hopefully) in a
standards-compliant way.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>