Cache the search paths and reuse them for subsequent library checks
This optimises the way libraries are searched for.  With a dynamic link,
the tag name is searched for independently of the library path.  This results
in approximately six times fewer attempted links.  For both dynamic and static
links, once a tag name and library path are found, the values are cached and
tried first when searching for subsequent libraries.  This results in a huge
speed increase for the second and later library checks.