boost.m4: provide a means to look for different library base names
A recent change seems to be dictated by the fact that on MingW, the
thread library is actually name thread_win32.  The patch is somewhat
uglified because the library name is also used to compute the name of
the output variables.  This shows that it is desirable to separate the
component name (e.g., thread), from the real shared-object base name
(e.g., thread or thread_win32).

* build-aux/boost.m4 (_BOOST_FIND_LIBS): New.
Includes the part of BOOST_FIND_LIB that used to look for the library,
inside the AC_CACHE_CHECK.
(BOOST_FIND_LIBS): New.
Contains what used to be around the library look up.
(BOOST_FIND_LIB): Backward compatibility layer.