boost.m4: do not issue useless -rpath /usr/local/lib
Serial 22.

Symptoms: on ArchLinux, some library linked with Boost thanks to
boost.m4 end up with the following errors:

vaucanson-git E: Insecure RPATH '/usr/local/lib' in file
('usr/lib/python3.4/site-packages/vcsn_cxx.so')
vaucanson-git E: Insecure RPATH '/usr/local/lib' in file
('usr/lib/vcsn/libvcsn.so.0.0.0')

The weird thing is that there is no reason for this dependency on
/usr/local/lib.  Careful reading of boost.m4 shows this is not
intended.  Indeed the empty library location ('') is tested first, but
the code that follows is not ready to handle an empty path, which
results in various errors that are misunderstood as meaning that the
libraries are not there.  So the next successful location tried (which
turns out to be /usr/local/lib) will be "accepted", and uselessly
propagated to the linker both as -L and -rpath.

* build-aux/boost.m4 (_BOOST_FIND_LIBS): Try first without any -L,
-rpath etc.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>