boost.m4: cope with GCC 5
Serial 24.

There are two needed adjustments.

One due to the output format which has changed:

  $ cat conftest.cc
  #include <boost/version.hpp>
  boost-lib-version = BOOST_LIB_VERSION
  $ g++-mp-4.9 -E conftest.cc            $ g++-mp-5 -E conftest.cc
  # 1 "conftest.cc"                      # 1 "conftest.cc"
  # 1 "<built-in>"                       # 1 "<built-in>"
  # 1 "<command-line>"                   # 1 "<command-line>"
  # 1 "conftest.cc"                      # 1 "conftest.cc"
  # 1 "/include/boost/version.hpp" 1 3   # 1 "/include/boost/version.hpp" 1 3
  # 2 "conftest.cc" 2                    # 2 "conftest.cc" 2
  boost-lib-version = "1_56"           ! boost-lib-version =
                                       ! # 2 "conftest.cc" 3
                                       !                    "1_56"

Since I fear that -P is not portable, let's play with tr and grep to
get rid of all the # lines and glue the remaining lines together.

The other one is that, finally, gcc 4.10 will actually be named gcc 5,
so recognize this name.

* build-aux/m4/boost.m4 (_BOOST_SED_CPP): Clean the preprocessor output
before handing it to the sed command.
(_BOOST_FIND_COMPILER_TAG): Support GCC 5.
Improve accuracy of error messages by adding delimiters to bad strings.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>