boost.m4: look for Python versioned libraries on Debian
I have been using this patch for my project since March 2015, and it
works nicely for Python 3.  I confess I don't have much experience
with Python 2.  But this was really needed for my project to work.

The idea is to look for the $PYTHON being used to decide which version
of Boost.Python should be used.

* build-aux/boost.m4 (BOOST_FIND_LIB, BOOST_FIND_LIBS, _BOOST_FIND_LIBS):
Take a lib suffix argument.
(Python): Specify the possible Python suffix version.