don't check for a library's existence if the tested ldpath = ''
If the tested ldpath = '', using it to create an absolute
path to a library leads to a path in the root directory, which
isn't what's intended. So don't.

Additionally if ldpath = '', it's equivalent to letting the linker
use only the system default paths, and there's no way of knowing in
which directory the linker might find the library.